Move exceptions.py to app/main/
